fix e2e
